+++ countingSort.h
@@ -3,5 +3,5 @@
 #include <cstring>
 using namespace std;
 
-void counting_sort_comp(int *arr, int capacity, int &count_comp);
+void counting_sort_comp(int *arr, int capacity, long long &count_comp);
 void counting_sort_time(int *arr, int capacity, long double &time);
