+++ selectionSort.h
@@ -3,5 +3,5 @@
 #include <iomanip>
 using namespace std;
 
-void selection_sort_comp(int *a, int n, int &count_comp);
-void selection_sort_time(int *a, int n,long double &time);
+void selection_sort_comp(int *a, int n, long long &count_comp);
+void selection_sort_time(int *a, int n, long double &time);
